chore: renaming file names for consistency
